Crate wayland_protocols_experimental

This crate provides Wayland object definitions for experimental protocol extensions.

This crate provides bindings for protocols that are officially evaluated, but not recommended for use outside of testing.

These bindings are built on top of the crates wayland-client and wayland-server.

Each protocol module contains a client and a server submodules, for each side of the protocol. The creation of these modules (and the dependency on the associated crate) is controlled by the two cargo features client and server.

Modules

input_method

This protocol allows applications to act as input methods for compositors.

keyboard_filter

This protocol allows applications to intercept and filter keyboard events.

session_management

The xx_session_manager protocol declares interfaces necessary to allow clients to restore toplevel state from previous executions.

text_input

This protocol allows applications to receive text composition events.