wayland_protocols_experimental::input_method::v1::client::xx_input_method_v1

Enum Request

#[non_exhaustive]
pub enum Request<'a> {
    PerformAction {
        action: WEnum<Action>,
    },
    CommitString {
        text: String,
    },
    SetPreeditString {
        text: String,
        cursor_begin: i32,
        cursor_end: i32,
    },
    DeleteSurroundingText {
        before_length: u32,
        after_length: u32,
    },
    MoveCursor {
        cursor: i32,
        anchor: i32,
    },
    Commit {
        serial: u32,
    },
    GetInputPopupSurface {
        surface: WlSurface,
        positioner: XxInputPopupPositionerV1,
    },
    Destroy,
}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

PerformAction

perform action

Perform an action on this text input.

Values set with this event are double-buffered. They must be applied and reset to initial on the next commit request.

The initial value of action is none.

Only available since version 3 of the interface

Fields

action: WEnum<Action>

action to perform

CommitString

commit string

Send the commit string text for insertion to the application.

Inserts a string at current cursor position (see commit event sequence). The string to commit could be either just a single character after a key press or the result of some composing.

The argument text is a buffer containing the string to insert. There is a maximum length of wayland messages, so text can not be longer than 4000 bytes.

Values set with this request are double-buffered. They must be applied and reset to initial on the next .commit request.

The initial value of text is an empty string.

Fields

text: String

SetPreeditString

pre-edit string

Send the pre-edit string text to the application text input.

Place a new composing text (pre-edit) at the current cursor position. Any previously set composing text must be removed. Any previously existing selected text must be removed. The cursor is moved to a new position within the preedit string.

The argument text is a buffer containing the preedit string. There is a maximum length of wayland messages, so text can not be longer than 4000 bytes.

The arguments cursor_begin and cursor_end are counted in bytes relative to the beginning of the submitted string buffer. Cursor should be hidden by the text input when both are equal to -1.

cursor_begin indicates the beginning of the cursor. cursor_end indicates the end of the cursor. It may be equal or different than cursor_begin.

Values set with this request are double-buffered. They must be applied on the next xx_input_method_v1.commit request. They must be reset to initial on the next committed disable event.

The initial value of text is an empty string. The initial value of cursor_begin, and cursor_end are both 0.

Fields

text: String

cursor_begin: i32

cursor_end: i32

DeleteSurroundingText

delete text

Remove the surrounding text.

before_length and after_length are the number of bytes before and after the current cursor index (excluding the preedit text) to delete.

If any preedit text is present, it is replaced with the cursor for the purpose of this event. In effect before_length is counted from the beginning of preedit text, and after_length from its end (see commit event sequence).

Values set with this request are double-buffered. They must be applied and reset to initial on the next xx_input_method_v1.commit request.

The initial values of both before_length and after_length are 0.

Fields

before_length: u32

after_length: u32

MoveCursor

move cursor and change selection

Unselects text, moves the cursor and selects text.

This is equivalent to dragging the mouse over some text: it deselects whatever might be currently selected and selects a new range of text.

The offsets used in arguments are in bytes relative to the current cursor position. Cursor is the new position of the cursor, and anchor is the opposite end of selection. If there’s no selection, anchor should be equal to cursor.

The offsets do not take preedit contents into account, nor is preedit changed in any way with this request.

Both cursor and anchor must fall on code point boundaries, otherwise text input client may ignore the request. It is therefore not recommended for an input method to move any of them beyond the text received in surrounding_text.

When surrounding_text is not supported, the offsets must not be interpreted as bytes, but as some human-readable unit at least as big as a code point, for example a grapheme.

The cursor and anchor arguments can also take the following special values: BEGINNING := 0x8000_0000 = i32::MIN END := 0x7fff_ffff = i32::MAX meaning, respectively, the beginning and the end of of all text in the input field.

Values set with this event are double-buffered. They must be applied and reset to initial on the next commit request.

The initial values of both cursor and anchor are 0.

Only available since version 3 of the interface

Fields

cursor: i32

anchor: i32

Commit

apply state

Apply state changes from commit_string, set_preedit_string and delete_surrounding_text requests.

The state relating to these events is double-buffered, and each one modifies the pending state. This request replaces the current state with the pending state.

The connected text input is expected to proceed by evaluating the changes in the following order:

Replace existing preedit string with the cursor.

Delete requested surrounding text.

Insert commit string with the cursor at its end.

Move the cursor and selection.

Calculate surrounding text to send.

Insert new preedit text in cursor position.

Place cursor inside preedit text.

Perform the requested action.

Note that the input method can not receive more than 4000 bytes of selection text, which might be the case for example when the entire document is selected. Nevertheless, the text input must delete the entire selected range before inserting the commit string.

The serial number reflects the last state of the xx_input_method_v1 object known to the client. The value of the serial argument must be equal to the number of done events already issued by that object. When the compositor receives a commit request with a serial different than the number of past done events, it must proceed as normal, except it should not change the current state of the xx_input_method_v1 object.

Fields

serial: u32

GetInputPopupSurface

create popup surface

Creates a new xx_input_popup_surface_v2 object wrapping a given surface.

The surface gets assigned the “input_popup” role. If the surface already has an assigned role, the compositor must issue a protocol error.

Issuing this request before receiving a committed .activate causes the “inactive” error.

Only available since version 2 of the interface

Fields

surface: WlSurface

positioner: XxInputPopupPositionerV1

Destroy

destroy the input method

Destroys the xx_input_method_v1 object and any associated child objects.

This is a destructor, once sent this object cannot be used any longer.

Implementations

impl<'a> Request<'a>

pub fn opcode(&self) -> u16

Get the opcode number of this message